Week9/day1/requirements.txt
pyautogen==0.2.35
openai
python-dotenv